🧹 clean up files for profiles types
+ public and private folders for profiles types
